# Conflicts:
#	xmipp3/constants.py
#	xmipp3/protocols/protocol_3dbionotes.py
#	xmipp3/protocols/protocol_align_volume.py
#	xmipp3/protocols/protocol_center_particles.py
#	xmipp3/protocols/protocol_deep_denoising.py
#	xmipp3/protocols/protocol_eliminate_empty_images.py
#	xmipp3/protocols/protocol_extract_particles.py
#	xmipp3/protocols/protocol_extract_unit_cell.py
#	xmipp3/protocols/protocol_movie_gain.py
#	xmipp3/protocols/protocol_pick_noise.py
#	xmipp3/protocols/protocol_preprocess/protocol_crop_resize.py
#	xmipp3/protocols/protocol_resolution3d.py
#	xmipp3/tests/test_protocol_extract_unit_cell.py
#	xmipp3/viewers/viewer_ctf_consensus.py
#	xmipp3/viewers/viewer_extract_unit_cell.py
#	xmipp3/viewers/viewer_movie_alignment.py

+++ xmipp3/protocols/nma/convert.py
@@ -24,11 +24,10 @@
 # *
 # **************************************************************************
 
-import os
 from collections import OrderedDict
 
 from pyworkflow.utils import Environ
-from pyworkflow.em.data import NormalMode
+from pwem.objects import NormalMode
 
 import xmippLib
 from xmipp3.convert import rowToObject, objectToRow
@@ -51,7 +50,7 @@ def rowToMode(row):
 
 def modeToRow(mode, row):
     """ Write the MetaData row from a given NormalMode object. """
-    row.setValue(xmippLib.MDL_ORDER, long(mode.getObjId()))
+    row.setValue(xmippLib.MDL_ORDER, int(mode.getObjId()))
     objectToRow(mode, row, MODE_DICT)
